brought developer only commands to developer.js
